requirements.txt
# Dependencies for Smartphone Price Prediction Dashboard

# High-Performance Data Engineering
polars>=0.20.0

# Data Extraction (Web Scraping)
beautifulsoup4>=4.12.0
requests>=2.31.0

# Data Version Control
dvc>=3.40.0

# Unit Testing & QA
pytest>=8.0.0

# Machine Learning Modeling (Phase 1)
# Machine Learning Modeling (Phase 1)
# Machine Learning Modeling (Phase 1)
scikit-learn>=1.3.0
rapidfuzz>=3.0.0
category_encoders>=2.6.0
pyarrow>=14.0.0
mlflow>=2.11.0
optuna>=3.6.0
xgboost>=2.0.0
torch>=2.2.0
chromadb>=0.4.0
langchain-huggingface>=0.1.0
sentence-transformers>=2.0.0
langchain-groq>=0.1.0
langchain-chroma>=0.1.0
python-dotenv>=1.0.0
